requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
idna==2.8
oauthlib==3.1.0
Pillow==5.4.1
pytz==2018.9
requests==2.22.0
urllib3==1.25.3
